Login working.
Dealing with the YML files is pretty simple, but hiding the
Github credentials was a bit strange. Figaro is meant for
Rails, so I had to parse the YML file like it were any other.

Up next is posting an issue from a bit of information in a
YML file created by an instructor.