Updated the test stories.
The main change here was to remove the text files since text
support is no longer expected. I also added a singular test
story as well as adding an additional repo to the test_stories.yml.
These two yml files should account for every case the storyteller is
responsible for.