Implemented Storyteller with tests.
The Storyteller class autoruns on initialize, making
the test suite burdensome on the requests per day.
Using Faraday, the connection the Storyteller makes
relies on the Github API and the basic_auth is enough
to gain access to the account. Important to note is
that label creation is only available to collaborators,
so each project will have to add the Github account
(ClassicRichard) as a collaborator before the stories
can be added correctly.